requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
idna==2.10
python-decouple==3.3
pytz==2020.1
request==1.0.117
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
